working migrated classifier with absolute imports
Changes to be committed:
	modified:   python_packages/profiles_classifier_pkg/MANIFEST.in
	deleted:    python_packages/profiles_classifier_pkg/profiles_classifier/Connector.py
	deleted:    python_packages/profiles_classifier_pkg/profiles_classifier/constants.py
	modified:   python_packages/profiles_classifier_pkg/profiles_classifier/model.py
	deleted:    python_packages/profiles_classifier_pkg/profiles_classifier/train.py
	new file:   python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/Connector.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/License.md -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/License.md
	new file:   python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/MLTrainer.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/README.md -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/README.md
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/RedshiftConnector.py -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/RedshiftConnector.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/SnowflakeConnector.py -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/SnowflakeConnector.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/MLTrainer.py -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/__init__.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/conda_env.yml -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/conda_env.yml
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/config/model_configs.yaml -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/config/model_configs.yaml
	new file:   python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/constants.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/logger.py -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/logger.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/predict.py -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/predict.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/requirements.txt -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/requirements.txt
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/tests/test_utils.py -> python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/tests/test_utils.py
	new file:   python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/train.py
	new file:   python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/utils.py
	deleted:    python_packages/profiles_classifier_pkg/profiles_classifier/utils.py
